pyproject.toml
[project]
name = "edu-assist"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]==0.135.3",
    "loguru==0.7.3",
    "openai==2.30.0",
    "pydantic==2.12.5",
    "python-dotenv==1.2.2",
    "pyyaml==6.0.3",
    "sympy==1.14.0",
]

[build-system]
requires = ["uv_build>=0.11.3,<0.12"]
build-backend = "uv_build"

[tool.fastapi]
entrypoint = "edu_assist.api:app"
